Code cleanup. Moved builders from shed to rocketStageBuilder. Removed boost lib from dockerFile (installation process was too long with this)

+++ core/rocket/fuelTank/FuelTank.h
@@ -27,7 +27,7 @@ public:
         return this->fuelAmount;
     }
 
-    [[nodiscard]] int getMaxEngineNumberSupport() const {
+    [[nodiscard]] size_t getMaxEngineNumberSupport() const {
         return this->maxEngineNumberSupport;
     }
